changed the DB changed the images store to s3 liara fixed many thing

+++ src/app/api/announcements/[id]/route.js
@@ -1,3 +1,4 @@
+import { getTrueImagesUrl, deleteImages } from '@/actions/image';
 import { getUser } from '@/lib/getUser';
 import prisma from '@/lib/prismaDB'
 import { NextResponse } from 'next/server'
@@ -13,12 +14,12 @@ export const GET = async (_req, { params }) => {
         }
 
         // Increment views by 1
-        const updatedAnnoncements = await prisma.announcements.update({
+        await prisma.announcements.update({
             where: { id },
             data: { views: announcements.views + 1 }
         });
 
-        return NextResponse.json(announcements);
+        return NextResponse.json(await getTrueImagesUrl(announcements));
     } catch (error) {
         return NextResponse.json({ message: `GET ANNOUNCEMENT ${id} ERROR`, error }, { status: 500 })
     }
@@ -31,6 +32,13 @@ export const DELETE = async (_req, { params }) => {
         const sessiion = await getUser()
         if (sessiion.user.role !== process.env.ADMIN_ROLE)
             return NextResponse.json({ message: "Unuthorized", error }, { status: 400 })
+        const announcement = await prisma.announcements.findUnique({ where: { id } });
+
+        if (!announcement) return NextResponse.json({ message: `ANNOUNCEMENT ${id} NOT FOUND` }, { status: 404 });
+
+        // Call deleteImages with the URLs of the images associated with the announcement
+        if (announcement?.imagesURL?.length > 0) await deleteImages(announcement.imagesURL);
+
         await prisma.announcements.delete({
             where: { id }
         })
@@ -64,7 +72,7 @@ export const PATCH = async (req, { params }) => {
         if (!updateAnnouncements) {
             return NextResponse.json({ message: `ANNOUNCEMENT ${id} NOT FOUND` }, { status: 404 })
         }
-        return NextResponse.json(updateAnnouncements);
+        return NextResponse.json(await getTrueImagesUrl(updateAnnouncements));
     } catch (error) {
         return NextResponse.json({ message: `PATCH ANNOUNCEMENT ${id} ERROR`, error }, { status: 500 })
     }
