+++ src/app/api/events/[id]/route.js
@@ -1,3 +1,4 @@
+import { deleteImages, getTrueImagesUrl } from '@/actions/image';
 import { getUser } from '@/lib/getUser';
 import prisma from '@/lib/prismaDB'
 import { NextResponse } from 'next/server'
@@ -12,11 +13,11 @@ export const GET = async (_req, { params }) => {
             return NextResponse.json({ message: `EVENT ${id} NOT FOUND` }, { status: 404 })
         }
         // Increment views by 1
-        const updatedEvents = await prisma.events.update({
+        await prisma.events.update({
             where: { id },
             data: { views: events.views + 1 }
         });
-        return NextResponse.json(events);
+        return NextResponse.json(await getTrueImagesUrl(events));
     } catch (error) {
         return NextResponse.json({ message: `GET EVENT ${id} ERROR`, error }, { status: 500 })
     }
@@ -29,6 +30,16 @@ export const DELETE = async (_req, { params }) => {
         const session = await getUser()
         if (session.user.role !== process.env.ADMIN_ROLE)
             return NextResponse.json({ message: "Unauthorized", error }, { status: 400 })
+        const events = await prisma.events.findUnique({
+            where: { id }
+        })
+        if (!events) {
+            return NextResponse.json({ message: `EVENT ${id} NOT FOUND` }, { status: 404 })
+        }
+
+        // Call deleteImages with the URLs of the images associated with the events
+        if (events?.imagesURL?.length > 0) await deleteImages(events.imagesURL);
+
         await prisma.events.delete({
             where: { id }
         })
@@ -63,7 +74,7 @@ export const PATCH = async (req, { params }) => {
         if (!updateEvents) {
             return NextResponse.json({ message: `EVENT ${id} NOT FOUND` }, { status: 404 })
         }
-        return NextResponse.json(updateEvents);
+        return NextResponse.json(await getTrueImagesUrl(updateEvents));
     } catch (error) {
         return NextResponse.json({ message: `PATCH EVENT ${id} ERROR`, error }, { status: 500 })
     }
