+++ src/app/api/news/[id]/route.js
@@ -1,3 +1,4 @@
+import { deleteImages, getTrueImagesUrl } from '@/actions/image';
 import { getUser } from '@/lib/getUser';
 import prisma from '@/lib/prismaDB'
 import { NextResponse } from 'next/server'
@@ -13,12 +14,12 @@ export const GET = async (_req, { params }) => {
         }
 
         // Increment views by 1
-        const updatedNews = await prisma.news.update({
+        await prisma.news.update({
             where: { id },
             data: { views: news.views + 1 }
         });
 
-        return NextResponse.json(news);
+        return NextResponse.json(await getTrueImagesUrl(news));
     } catch (error) {
         return NextResponse.json({ message: `GET NEWS ${id} ERROR`, error }, { status: 500 })
     }
@@ -31,6 +32,17 @@ export const DELETE = async (_req, { params }) => {
         const sessiion = await getUser()
         if (sessiion.user.role !== process.env.ADMIN_ROLE)
             return NextResponse.json({ message: "Unuthorized", error }, { status: 400 })
+
+        const news = await prisma.news.findUnique({
+            where: { id }
+        })
+        if (!news) {
+            return NextResponse.json({ message: `NEWS ${id} NOT FOUND` }, { status: 404 })
+        }
+
+        // Call deleteImages with the URLs of the images associated with the news
+        if (news?.imagesURL?.length > 0) await deleteImages(news.imagesURL);
+
         await prisma.news.delete({
             where: { id }
         })
@@ -64,7 +76,7 @@ export const PATCH = async (req, { params }) => {
         if (!updateNews) {
             return NextResponse.json({ message: `NEWS ${id} NOT FOUND` }, { status: 404 })
         }
-        return NextResponse.json(updateNews);
+        return NextResponse.json(await getTrueImagesUrl(updateNews));
     } catch (error) {
         return NextResponse.json({ message: `PATCH NEWS ${id} ERROR`, error }, { status: 500 })
     }
